+++ source_gen/test/builder_test.dart
@@ -556,12 +556,93 @@ part "a.foo.dart";''',
         const CombiningBuilder(),
         {
           '$_pkgName|lib/a.dart': 'library a; part "a.g.dart";',
+          '$_pkgName|lib/a.foo.g.part': 'class A {}',
+        },
+        onLog: _failOnSevereLog,
+        outputs: {
+          '$_pkgName|lib/a.g.dart': decodedMatches(
+            startsWith('// GENERATED CODE - DO NOT MODIFY BY HAND'),
+          ),
+        },
+      );
+    });
+
+    test('throws with bad code', () async {
+      final logs = <String>[];
+      await testBuilder(
+        const CombiningBuilder(),
+        {
+          '$_pkgName|lib/a.dart': 'library a; part "a.g.dart";',
           '$_pkgName|lib/a.foo.g.part': 'some generated content',
+        },
+        onLog: (log) {
+          logs.add(log.message);
         },
         outputs: {
           '$_pkgName|lib/a.g.dart': decodedMatches(
             startsWith('// GENERATED CODE - DO NOT MODIFY BY HAND'),
           ),
+        },
+      );
+      expect(
+        logs,
+        contains(
+          contains(
+            'An error `FormatterException` occurred while formatting the ',
+          ),
+        ),
+      );
+    });
+
+    test('uses a custom header when provided', () async {
+      await testBuilder(
+        const CombiningBuilder(header: _customHeader),
+        {
+          '$_pkgName|lib/a.dart': 'library a; part "a.g.dart";',
+          '$_pkgName|lib/a.foo.g.part': 'class C {}',
+        },
+        onLog: _failOnSevereLog,
+
+        outputs: {
+          '$_pkgName|lib/a.g.dart': decodedMatches('''
+$_customHeader
+
+part of 'a.dart';
+
+class C {}
+'''),
+        },
+      );
+    });
+
+    test('includes no header when `header` is empty', () async {
+      await testBuilder(
+        const CombiningBuilder(header: ''),
+        {
+          '$_pkgName|lib/a.dart': 'library a; part "a.g.dart";',
+          '$_pkgName|lib/a.foo.g.part': 'class A {}',
+        },
+        onLog: _failOnSevereLog,
+        outputs: {
+          '$_pkgName|lib/a.g.dart': decodedMatches('''
+part of 'a.dart';
+
+class A {}
+'''),
+        },
+      );
+    });
+
+    test('formats the output content', () async {
+      await testBuilder(
+        const CombiningBuilder(),
+        {
+          '$_pkgName|lib/a.dart': 'library a; part "a.g.dart";',
+          '$_pkgName|lib/a.foo.g.part': '\n\n  class   A   {}  \n\n',
+        },
+        onLog: _failOnSevereLog,
+        outputs: {
+          '$_pkgName|lib/a.g.dart': decodedMatches(contains('\nclass A {}\n')),
         },
       );
     });
@@ -941,10 +1022,7 @@ foo generated content
           contains('// "int x" is deprecated!'),
         ),
       },
-      onLog: (log) {
-        if (log.level < Level.SEVERE) return;
-        fail('Unexpected log message: ${log.message}');
-      },
+      onLog: _failOnSevereLog,
     );
   });
 }
@@ -959,10 +1037,7 @@ Future<void> _generateTest(CommentGenerator gen, String expectedContent) async {
     outputs: {
       '$_pkgName|lib/test_lib.foo.dart': decodedMatches(expectedContent),
     },
-    onLog: (log) {
-      if (log.level < Level.SEVERE) return;
-      fail('Unexpected log message: ${log.message}');
-    },
+    onLog: _failOnSevereLog,
   );
 }
 
@@ -1136,3 +1211,10 @@ part of 'test_lib.dart';
 
 // hello
 ''';
+
+void _failOnSevereLog(LogRecord log) {
+  if (log.level < Level.SEVERE) return;
+  addTearDown(() {
+    fail('Unexpected log message: ${log.message}');
+  });
+}