+++ source_gen/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.2.0
+
+- Add `header` configuration to `CombiningBuilder`.
+- `CombiningBuilder` now formats the generated code with the `DartFormatter`.
+
 ## 4.1.2
 
 - Allow `analyzer: '>=8.1.1 <11.0.0'`.